Refactoring: move funcs to modules
